{{cookiecutter.directory_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.directory_name}}"
version = "0.1.0"
description = ""
authors = ["{{cookiecutter.author_name}}"]


[tool.poetry.dependencies]
python = "^3.8"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
notebook = "^6.4.10"
pre-commit = "^2.19.0"
gitpython = "^3.1.27"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''